@@ -4,7 +4,8 @@ buildscript {
         google()
     }
     dependencies {
-        classpath 'com.android.tools.build:gradle:3.0.0'
+        classpath 'com.android.tools.build:gradle:3.6.1'
+
 
         // NOTE: Do not place your application dependencies here; they belong
         // in the individual module build.gradle files
@@ -26,22 +27,22 @@ task clean(type: Delete) {
 ext {
     // Sdk and tools
     // Support library and architecture components support minSdk 14 and above.
-    minSdkVersion = 14
-    targetSdkVersion = 26
-    compileSdkVersion = 26
-    buildToolsVersion = '26.0.2'
+    minSdkVersion = 21
+    targetSdkVersion = 28
+    compileSdkVersion = 28
+
 
     // App dependencies
-    supportLibraryVersion = '26.1.0'
-    guavaVersion = '18.0'
-    junitVersion = '4.12'
+    supportLibraryVersion = '28.0.0'
+    junitVersion = '4.13'
     mockitoVersion = '1.10.19'
     powerMockito = '1.6.2'
     hamcrestVersion = '1.3'
-    runnerVersion = '1.0.1'
-    rulesVersion = '1.0.1'
-    espressoVersion = '3.0.1'
-    roomVersion = "1.0.0"
+    runnerVersion = '1.0.2'
+    rulesVersion = '1.0.2'
+    espressoVersion = '3.2.0-beta01'
+    roomVersion = '2.1.0'
+    stethoVersion = '1.5.1'
 }
 
 /*
