@@ -17,9 +17,10 @@
 package com.example.android.architecture.blueprints.todoapp.addedittask;
 
 import android.content.Context;
-import android.databinding.ObservableBoolean;
-import android.databinding.ObservableField;
-import android.support.annotation.Nullable;
+
+import androidx.annotation.Nullable;
+import androidx.databinding.ObservableBoolean;
+import androidx.databinding.ObservableField;
 
 import com.example.android.architecture.blueprints.todoapp.R;
 import com.example.android.architecture.blueprints.todoapp.data.Task;
@@ -30,7 +31,7 @@ import com.example.android.architecture.blueprints.todoapp.data.source.TasksRepo
  * ViewModel for the Add/Edit screen.
  * <p>
  * This ViewModel only exposes {@link ObservableField}s, so it doesn't need to extend
- * {@link android.databinding.BaseObservable} and updates are notified automatically. See
+ * {@link androidx.databinding.BaseObservable} and updates are notified automatically. See
  * {@link com.example.android.architecture.blueprints.todoapp.statistics.StatisticsViewModel} for
  * how to deal with more complex scenarios.
  */
