@@ -16,19 +16,20 @@
 
 package com.example.android.architecture.blueprints.todoapp;
 
-import static android.support.test.InstrumentationRegistry.getInstrumentation;
-import static android.support.test.runner.lifecycle.Stage.RESUMED;
-
 import android.app.Activity;
 import android.content.pm.ActivityInfo;
 import android.content.res.Configuration;
-import android.support.annotation.IdRes;
-import android.support.annotation.NonNull;
 import android.support.test.runner.lifecycle.ActivityLifecycleMonitor;
 import android.support.test.runner.lifecycle.ActivityLifecycleMonitorRegistry;
 import android.support.v7.widget.Toolbar;
 
+import androidx.annotation.IdRes;
+import androidx.annotation.NonNull;
+
 import java.util.Collection;
+
+import static android.support.test.InstrumentationRegistry.getInstrumentation;
+import static android.support.test.runner.lifecycle.Stage.RESUMED;
 
 /**
  * Useful test methods common to all activities
@@ -63,7 +64,7 @@ public class TestUtils {
      */
     public static String getToolbarNavigationContentDescription(
             @NonNull Activity activity, @IdRes int toolbar1) {
-        Toolbar toolbar = (Toolbar) activity.findViewById(toolbar1);
+        Toolbar toolbar = activity.findViewById(toolbar1);
         if (toolbar != null) {
             return (String) toolbar.getNavigationContentDescription();
         } else {
